# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "Swagger Example API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/product": {
            "get": {
                "description": "Method to search product",
                "parameters": [
                    {
                        "type": "string",
                        "description": "product name",
                        "name": "name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/qualification_internal_usecase_model.Product"
                        }
                    }
                }
            },
            "post": {
                "description": "Method to create product",
                "parameters": [
                    {
                        "description": "Body",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/qualification_internal_usecase_model.Product"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/product/{id}": {
            "get": {
                "description": "Method to get product",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "product id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/qualification_internal_usecase_model.Product"
                        }
                    }
                }
            }
        },
        "/products": {
            "get": {
                "description": "Get All products",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/qualification_internal_usecase_model.Product"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "qualification_internal_usecase_model.Product": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    }
}